Prepare to run multiple tests after each other.
The current setup makes it easy to run one test on one server. Thats'
not the best way to use resources. We should aim for running multiple
tests after each other.

This is the first step to reach that goal by renaming server -> test.
The next step is to make it configurable to run multiple tests.

Bug: T225416
Change-Id: I76bb9b11083057f6c27690decde21646027b2a93