Run test on native connection for emulated mobile
Test running tests using native connection for emulated mobile
to see if we still see the big stdev. This isn't perfect (who knows
whatever connection we have on the machine) but we can at least try it
and see if we see the same pattern.

Let this run over the weekend (and longer) if needed and then we can just
remove the new config and the change in run.

Bug: T233789
Change-Id: Ibee12f5caac61ab9a8adec586c0eb4592838d090