Disable WebDriver interruptions
After discussion with Mozillas performance team it looks like one way
to get more stable metrics is to try reduce JavaScript calls to the
browser during the load of the page.

With this configuration change, we let the driver start the navigation then
leave the browser do its stuff until X seconds, and then we check if
its ready every Y seconds.

This also disables the loading of the Browsertime Extension for Firefox
to try limit the things in the browser.

Bug: T239780
Change-Id: I2a38fc891a504f50b46bc333a24a7b17ccc7067c