Disable safe browsing for Firefox
One known problem with Firefox in perfromance testing, is that on the first loadit starts to connect and download lists to update the safe browsing lists.
This happens at the same time as we run our tests. It was disabled but I think
that in 69 things changed (we had the same problem in WebPageTest).

This is fixed upstream in Browsertime but we can try it out like this first.

Bug: T239780
Change-Id: Ibb7eed78e3cb2c18e351b1b38d85b7af5e71302c