Remove container if there's a update in the project.
One problem is that we run out of space of the agent when we have
multiple containers downloaded on the server. With this fix we
remove the container if there's an change in git. That we we will
download the container more often than before but we only do it
on a change in git. And that we will only have one instance of the
container on the server at the same time.

Bug: T240095
Change-Id: Ib2a6ad628f4fa34684c7e8fbe10109db40ef85bf