Use minimal container for WebPageTest tests.
This starts to use the slim debian based container for the WebPageTest
tests. They run on their own machine and do not do Browsertime tests so
we can just use the slim one.

We also upgrade the other machines to use the latest version (that also
removes a lot of extra kb).

Bug: T247164
Change-Id: Iabb0f2f3e43f8882569c26720241e396b9315000