Revise legacyObama search
A crude implementation for legacy search was created in
I2dcf5d81821d4b2ba18f45f05648124edac89b4e, however it proved to be too
crude to be useful. This commit instead relies on user timings performed
by the client in an attempt to get much more accurate metrics.

Additionally, since we need to compare Vue search with legacy search, a
`searchScriptFactory` was created that makes scripts that only vary by
several params. With this factory, vue search and legacy search will be
able to be kept DRY. The Vue script will be added in a follow-up patch.

Note to reviewers: The test in this patch relies on a core
patch (see Depends-On). In order to test the script and the associated
core patch locally you will need to change the URL variable in each
script to point to your host machine (on Mac this can be accomplished
with `http://host.docker.internal:PORT`) as well as create the following
article pages:

* "B",
* "Bananabird",
* "Banana"

Depends-On: I0fa6b8904bd43c87a68e9161f00d686a0e588966
Bug: T251544
Change-Id: I732690099032af95a1f2cf5ae74cd7abbd473438