Add Wvui search script and point legacy/Wvui search tests to beta
Adds a test and config file for the new Wvui search component. This
commit also amends the searchScriptFactory to handle client-side
rendering of the Wvui search component (Vector's server-rendered search
input is replaced when the Wvui search finishes loading).

Additionally, since Vue search has now been enabled on beta (and legacy
search was already enabled), we can temporarily point the tests to those
endpoints (with the plan to switch to production when deployed there).

Bug: T251544
Depends-On: I14e44b45a66213821d69cd22395fedbae747da88
Change-Id: I73c2157d070545d0fa6c9e7def5d88f98afb2f4a