Point Vue/Legacy search tests to frwiki and update query string
These tests were pointing at the beta cluster, but for some reason the
beta cluster is currently only showing exact title matches and, as a
result, the dashboard has stopped displaying results [1]. Regardless of
the issue there, testing a production wiki is better.

* Also change search query to test "A", "Abc", "Ab" instead of "Banana"
as the former has a higher chance of showing suggestions.
* Additionally refactors Xpath command into its own function and adds
quotes around `highlightClass` (correcting a bug)

[1] https://grafana.wikimedia.org/d/GivPpdsGk/vue-vs-legacy-search?orgId=1

Bug: T251544
Change-Id: I1aa4eac0239c1040868b6156b2e0c4b690f54418