Set CPU throttle to 1 to avoid upgrade problems.
It's been a bug in Chrome for a long time that the CPU throttling
hasn't worked. See https://github.com/sitespeedio/browsertime/issues/1380

However last night I found a workaround that fixes it but that will affect
our metrics, so before I upgrade with the fix, lets turn it off and then
in a more controlled way turn it on and adjust the alerts.

Change-Id: I521c1d5ac0a6dc220311a2361ae223609a12745f