Disable Chrome PaintHoldingCrossOrigin feature.
See https://bugs.chromium.org/p/chromium/issues/detail?id=1291502#c22

Also want to disable for emulated mobile but the CPU throttling is
already using the disable features functionality since a bug in Chrome
and disabling twice will break it. I need to do a fix in Browsertime
for that.

Bug: T299886
Change-Id: I9b78db3d733cf436544b9a8e6bf5dbb848d06835