Enable PaintHoldingCrossOrigin for investigating LCP changes.
In our bug report to Google we only reported that FCP was affected,
however in my tests I could see that also LCP is affected. Lets
enble PaintHoding for a couple of days and compare LCP from Chrome
vs video and get data for an even better upcoming blog post.

Bug: T299886
Change-Id: I5e9c81d26094be106c9567d1c96ee6613875d4a1