Increase number of runs for instant tests.
With the new setup where the instant tests runs on a new machine we can run more iterations.
Lets do 21 for desktop and 7 for emulated mobile (we can increase emulated if needed).

Also remove a couple of URLs that we do not use and add the main page for desktop tests.

Bug: T346637
Change-Id: I9045fc0abafbef96d3bd7731950038b04d8cf8d9