Test dark mode on a dedicated server.
These are the tests I've been running locally for Chrome, lets
run it on the dedicated host and dive into the metrics there.

This will fail hard on Firefox but I think that is ok for now.

For future improvements on the tests, I need to fix so we can
measure user journeys using WebPageReplay. At the moment one
view tests is kind of hard coded in the scripts. Let me create
a task for fixing that. For our case that would had helped us
to use the dedicated bare metal servers.

Bug: T349308
Change-Id: I750d9aa073931a210283981830bcb1513199325d