Rename webpagereplay tests, unify latency and test it.
There's a couple of changes here, they are done in one go so its easier
to deploy them (each servers tests needs to be stopped and updated).

We rename the WebPageReplay test folders to make it easier to understand
what tests runs where. And then we remove setting different latency for
different tests, we just follow Mozilla and run the tests as fast as we can.

And then as a last thing we change group 1 tests to test on Italian wiki,
since that content is larger = makes it easier to spot regressions.

Also now run the emulated mobile tests on one server, making it more easy
to remove them when we gets real devices.

We also start by rebaselining all the tests.

Bug: T356839
Change-Id: Ib21e05b7b30d156c715fd6fc9909f6ac18b2b1e2