Run stable release and upgrade Chrome to 130 and Firefox to 131.
Instead of using pinned exact releases we will now try to run on
major releases. This way, the containers can get updated on the
server without us doing a commit.

We still prune the container but let us switch to pull in a couple
of weeks.

This change also upgrades from Chrome 129 -> 130 and Firefox
130 -> 131.

Bug: T377414
Change-Id: Ib93317f00e8051e6e327b222e5a91e11d4f5370e